+++ src/App.jsx
@@ -1,16 +1,8 @@
 // tradimedika-v1/src/App.jsx
-import Hero from "./components/sections/Hero";
-import Footer from "./layout/Footer";
-import Header from "./layout/Header";
+import Router from "./routes/Router";
 
 function App() {
-  return (
-    <div className="bg-light dark:bg-dark flex h-screen flex-col items-center justify-between transition duration-300 ease-in-out">
-      <Header />
-      <Hero />
-      <Footer />
-    </div>
-  );
+  return <Router />;
 }
 
 export default App;